3gpp channel: fix a bug "IsConnected" (add IsValidLink check instead of IsConnected) and extend 3gpp channel unit test
Currently 3gpp channel supports calculation of only UE<->gNb channel, but not UE<->UE and gNB<->gNb. IsBeamforming was
including wrongly the call to IsConnected which was returning true if the devices were connected. So, the beamforming
interference was not being calculated between UEs and gNbs that were not associated. Now the function IsValidLink is called
to check if the link is UE<->gNb no matter if they are connected or not. 3gpp channel unit test is extended to account
for IsValidLinkFunction.