fix example: cttc-3gpp-channel-simple-fdm was not running properly after changes in EPC and mmwave-helper
Added proper activation of dedicated bearer. Function that sends the packet and the RLC and PDCP connectors
are update to connect to this dedicated bearer.

Added condition to check if RLC and PDCP callbacks have run. If any of these did not execute the main will
return FAILURE. This shall help early detection if the example breaks again in future.