nr: Removed UL/DL methods
Having different methods for UL or DL transmission can lead
to the proliferation of the public API methods. Hence,
cut them by having a single "SetDl" or "SetUl" method,
that would setup everything needed to work for DL or for
UL.

The call to the Set method is done inside the scheduler, which
is the only source of knowledge for the AMC objects.