NrUeMac: Rework the logic to TX SL PSCCH and PSSCH
The reasons of this reworking are:

1. The code was hitting several asserts when there
is a grant for a slot and the RLC buffer is empty.
In such case, DoTransmitNrSlRlcPdu will not be called
and the HARQ entity will not receive any packet,
and then at the time of retransmission HARQ entity
would return an empty packet burst.

2. slResoReselCounter and cReselCounter should be
decreased after the last transmission of the MAC
PDU as per 38.321 5.22.1.3.1a. Before, we were
decreasing only after the first transmission, which
was creating issue upon re-selection. That is, when
slResoReselCounter goes to zero after first retransmission,
and we need to re-select the resources, the old grants
must be erased. Doing so will also erased the grants
for the retransmission (s) of the TB in the HARQ buffer.