nr-spectrum-value-helper: unify functions for spectrum creation
Motivation is that allowing two different functions does not guarantee
us uniqness in the way of creating the spectrum. Basically, it could
cause in future bugs, and hence is better to have only one way of
creating it.
Now pass SCS instead of numerology, rbNum instead of BW where used
the old function.