realistic-beamforming test: relax tolerance for QUICK mode
With previous configuration of tolerance, in QUICK mode is like
if tolerance did not exist. Because if 3/4 test configurations
give expected results, that is fine. But that gives 0.75, and
if test tolerance is 0.2, that test will fail, even if it is
tolerable error 1 of 4, and it should pass. So with this fix we
increase the tolerance for the QUICK mode.