ftp/utils: Added FTP Model 1 Helper to nr module and ftp util files
FTP model 1 helper is not NR specific helper, could be used by another
modules, but it still needs to be developed, hence for the moment it will
stay in nr module.

Added files that enable file transfer applications. These files are
distributed with GPL v2.0 licensed, and as such are added, with
an idea and goal to be improved during its usage and then proposed
to be included into ns-3-dev/applications module.