nr rem: (fixes #121) Add command line parameters, doxygen, fix case when simTag empty
Add command line parameters that facilitate usage of rem-example.cc. Add doxygen
to rem-example.cc and rem-beam-example.cc with more specific info on how to plot the
rems once that the simulation is finished.

Fix in REM helper the case when simTag is empty, for some reason was disabled
generation of REM figures.