srs scheduling: dont assert in the case that not all UEs received RAR message
Previously the code was asserting if the UE with the RNTI was not found
in the list of UEs. In some cases this assert was not being fullfilled
because UE did not receive RAR message and did not set its RNTI. This happens
currently around 20ms from the beginning of the attachment.