Consider FREQ domain for sensing based resource selection and resource allocation
Now, the sensing-based resource selection considers subchannel overlapping to
discard a slot. Moreover, at the scheduler irrespective of sensing or no-sensing
the scheduler is able to randomly select the required number of contiguous subchannels
among the available ones. SCI 1-A, header now can indicate the starting subchannel
index of the possible retransmissions of a TB.