PktTxRx,V2X examples: Sum the pkt and header size to correctly compute the real pkt size
TxWithSeqTsSize and RxWithSeqTsSize trace sources contain the Ptr to the pkt
that does not has the SeqTs header. However, in case of TxWithSeqTsSize we
can not store the Ptr to the pkt to retrieve the pkt size later because the
application adds SeqTs header right after the trace source gets triggered.