gitlab-ci: Limit the tests to "nr" module and run all examples
Currently it is not possible to limit the tests/examples to a specific module.
However, for tests it is possible to pass the asterix, and then if tests have the same prefix like in the NR module, we can limit the scop of the tests to be run to these tests.
On the other hand for the examples it is not yet possible to limit the scope in a simillar way.

This should reduce a little bit job execution time, so hopefully it does not exceeds 3 hour limitation by gitlab.