tests: Fix HARQ timings in nr-lte-pattern-generation
The test was not taking into account S slots as the opportunity to transmit
DL HARQ feedback.

After fixing a bug in the NR module that was not allowing the S slots to be used
for UL CTRL, this issue in the test poped up.

Explanation of testing DL harq feedback timing:

E.g. {0, 6} means that for the first DL slot from the pattern the HARQ feedback
from the UE can be sent earliest after 4 slots (n1 delay),
but 4th and 5th slots are the DL slots, so UE cannot use any
of these slots for to transmit HARQ feedback, so it has to wait
6 slots (6th slot is the special slot, which has the uplink CTRL).
Hence having the pattern:
      DL,0
      S, 1 -> DL HARQ fb can be sent after 5 slots in the S slot
      UL,2 -> /
      UL,3 -> /
      DL,4 -> DL HARQ fb can be sent after 4 slots in the UL slot
      DL,5 -> DL HARQ fb can be sent after 6 slots which is the S slot
      S, 6 -> DL HARQ fb can be sent after 5 slots which is the S slot
      UL,7 -> /
      UL,8 -> /
      DL,9 -> DL HARQ fb can be sent after 4 slots in the UL slot