Revert "nr-phy, nr-gnb-phy, nr-spectrum-phy: Resolving issue detected in NR-U simulations"
This reverts commit a4438bc40e029282320c92bb2ef42fe3952e061a.

Merge request that adds the reproducibility !37 testing has detected an issue with recently merged commit #a4438bc4 from !29 (merged) MR.
We need to revert back this #a4438bc4 commit until the issue is resolved. The issue is that after this commit, the uplink flows in cttc-3gpp-channel-nums-fdm are 0.

See issue 149.